refactor: create separate package for services, repositories, models and routers
