+++ app/models/request/refresh.py
@@ -1,0 +1,7 @@
+from pydantic import constr
+
+from app.models.models import CamelModel
+
+
+class RefreshSchema(CamelModel):
+    refresh_token: constr()
