+++ app/models/token.py
@@ -1,0 +1,5 @@
+from app.models.models import CamelModel
+
+
+class Token(CamelModel):
+    token: str
